requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
requests==2.31.0
python-dotenv==1.0.1
